requirements.txt
numpy
pandas
matplotlib
fastapi
uvicorn
pydantic
pyyaml
pytest
